use null stream instead of /dev/null
When HEAD fails, a GET is performed. To avoid keeping the
potentially DVD sized data in memory or on disk, it was
piped to /dev/null. This doesn't work on Microsoft Windows
where there is no /dev/null. There is a NUL device, but
I had trouble testing it and I didn't want to add a special
case for 1 platform. Instead I created a null stream, a
blackhole for data to go. This should be more portable
and more performant.

Thanks to @4lexBaum for the original bug report:
https://github.com/tcort/markdown-link-check/issues/1