Replace isemail, convert to ES6 module
Replace isemail with node-email-verifier, which also checks that email
domains have an MX record.

Add support for the timeout option for domain MX verification.

This also fixes the deprecation warning for the punycode module, which
is a dependency of isemail.

Convert the module to ES6 with `"type": "module"` in package.json. This
will mean that users will need to import into an ES6 module or use one
of the workarounds for cjs such as `await import`.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>